double (64bit) instead of float datatype (32bit)
 I'm not sure of any situation in which memory was so tight that one had to sacrifice accuracy for space

+++ FCFS.java
@@ -5,7 +5,7 @@ public class FCFS {
     public static void main(String args[]) throws Exception {
 
         int n; // Number of processes
-        float averageWaitTime = 0; // Avg waiting time
+        double averageWaitTime = 0; // Avg waiting time
         InputStreamReader isr = new InputStreamReader(System.in);
         BufferedReader br = new BufferedReader(isr);
